chore: move type files up and cleanup exports
