+++ src/common/ar-io.ts
@@ -22,8 +22,8 @@ import {
   EvaluationParameters,
   Gateway,
   SmartWeaveContract,
-} from '../types/index.js';
-import { RemoteContract } from './index.js';
+} from '../types.js';
+import { RemoteContract } from './contracts/remote-contract.js';
 
 // TODO: append this with other configuration options (e.g. local vs. remote evaluation)
 export type ContractConfiguration =
