+++ src/common/index.ts
@@ -15,6 +15,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 export * from './ar-io.js';
-export * from './contracts/index.js';
 export * from './error.js';
 export * from './logger.js';
+
+// contracts
+export * from './contracts/remote-contract.js';
+export * from './contracts/warp-contract.js';
