+++ src/common/logger.ts
@@ -18,7 +18,7 @@
 import 'setimmediate';
 import winston, { createLogger, format, transports } from 'winston';
 
-import { Logger } from '../types/index.js';
+import { Logger } from '../types.js';
 import { version } from '../version.js';
 
 export class DefaultLogger implements Logger {
