+++ src/contract-state.ts
