+++ src/node/index.ts
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-export * from '../types/index.js';
+export * from '../types.js';
 export * from '../common/index.js';
 export * from '../constants.js';
 export * from '../utils/index.js';
