+++ src/types.ts
@@ -14,7 +14,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 export * from './arns-service.js';
 export * from './contract-state.js';
 export * from './common.js';
