+++ tsconfig.json
@@ -24,6 +24,6 @@
     "skipLibCheck": true,
     "strictNullChecks": true
   },
-  "include": ["src", "tests/utils/arweave.test.ts"],
-  "exclude": ["lib", "node_modules", "bundles"]
+  "include": ["src"],
+  "exclude": ["lib", "node_modules", "bundles", "tests"]
 }
